VMO-6755 move programmatic logic in import store
